Add files via upload
Election Exit Poll Prediction and U.S.A Presidential Speech Analysis using Machine Learning -
The first project is to predict which party a citizen is going to vote for on the basis of their age and according to the answers given by the citizens to the questions asked in a survey conducted. The second project is based on the analysis of the inaugural U.S.A. Presidential speeches. One has to draw inferences based on the analysis done on these speeches.